chore: move updater to its own folder and create utils file
